Actually added screen scrolling
the screen scrolls, and generates 1 out of 3 maps randomly. Collision is still broke.

+++ BombsMain/Blocks.pde
@@ -2,7 +2,7 @@
 
 //Muren, moet nog collision op
 class Wall extends Object {
-  Wall(int x, int y, int w, int h, ObjectHandler objectHandler, Sprites sprites) {
+  Wall(float x, float y, int w, int h, ObjectHandler objectHandler, Sprites sprites) {
     super(x, y, w, h, ObjectID.WALL, objectHandler, sprites);
   }
 
@@ -20,7 +20,7 @@ class Wall extends Object {
 
 //Breekbare blocks, alles moet nog 
 class BreakableBlocks extends Object {
-  BreakableBlocks(int x, int y, int w, int h, ObjectHandler objectHandler, Sprites sprites) {
+  BreakableBlocks(float x, float y, int w, int h, ObjectHandler objectHandler, Sprites sprites) {
     super(x, y, w, h, ObjectID.BBLOCKS, objectHandler, sprites);
   }
 
@@ -36,7 +36,7 @@ class BreakableBlocks extends Object {
 
 //Onder en boven muren
 class Rock extends Object {
-  Rock(int x, int y, int w, int h, ObjectHandler objectHandler, Sprites sprites) {
+  Rock(float x, float y, int w, int h, ObjectHandler objectHandler, Sprites sprites) {
     super(x, y, w, h, ObjectID.ROCK, objectHandler, sprites);
   }
